% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_qrf.R
\name{partial_qrf}
\alias{partial_qrf}
\title{Partial Dependence Plot for Quantile Random Forest}
\usage{
partial_qrf(object, pred.var, Q = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{object}{a \code{\link[ranger]{ranger}} quantile random forest object.}

\item{pred.var}{character string giving the names of the predictor variables of
interest (see \code{\link[pdp]{partial}}).}

\item{Q}{a numeric vector of probabilities for which the plot is desired.}

\item{...}{other arguments passed to \code{\link[pdp]{partial}}.}
}
\value{
A \code{data.frame} with the following columns:
the predictor variables supplied in \code{pred.var},
response variable (the name is extracted from the \code{ranger} function call),
and, if \code{length(Q) > 1}, one more column named \code{"Quantile"}.
}
\description{
Get a data frame for partial dependence plots from a quantile random forest.
The plots can be obtained with a variety of plotting functions from other packages.
}
\examples{
\dontrun{
# Example 1: Swiss
library(dplyr)
library(ggplot2)

# fit a quantile random forest
qrf <- ranger::ranger(Examination ~ ., data = swiss, quantreg = TRUE, num.trees = 50)

# a plot for one quantile
partial_qrf(qrf, pred.var = "Agriculture", Q = 0.5) \%>\%
    ggplot(aes(Agriculture, Examination)) +
    geom_line()

# a plot for several quantiles, with decorations
library(hrbrthemes)
library(viridis)
partial_qrf(qrf, pred.var = "Agriculture") \%>\%
    ggplot(aes(Agriculture, Examination, group = Quantile, color = Quantile)) +
    geom_line() +
    scale_color_viridis(discrete = TRUE) +
    theme_ipsum()

# a plot for one quantile for 2 predictors (takes longer, so save as an object)
df <- partial_qrf(qrf, pred.var = c("Agriculture", "Catholic"), Q = 0.5)
ggplot(df, aes(Agriculture, Catholic)) +
    geom_tile(aes(fill = Examination)) +
    scale_fill_viridis() +
    theme_ipsum() +
    labs(fill = "Median\nexamination")
}

}
\seealso{
\code{\link[pdp]{partial}}, \code{\link[ranger]{ranger}}
}
\keyword{forest}
\keyword{quantile}
